🚚 Rename: 예약 관련 API를 reservation 도메인으로 이동
