+++ src/features/reservation/activity-panel/mutations/useCreateReservation.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { createReservation } from '@/features/activity/apis';
+import { createReservation } from '@/features/reservation/apis';
 
 export const useCreateReservation = (activityId: number) => {
   return useMutation({
