+++ src/features/reservation/apis.ts
@@ -1,5 +1,8 @@
+import { z } from 'zod';
+import { ActivityScheduleSchema } from '@/features/activity/types';
 import type {
   CancelMyReservationRequestBody,
+  CreateActivityReservationRequestBody,
   CreateMyReservationReviewRequestBody,
   GetMyReservationsQuery,
   MyReservationItem,
@@ -11,6 +14,7 @@ import {
   GetMyReservationsResponseSchema,
 } from '@/features/reservation/types';
 import { bffFetch } from '@/shared/apis/base/bffFetch';
+import { get } from '@/shared/apis/base/publicFetch';
 
 export const getMyReservations = async (query?: GetMyReservationsQuery) => {
   const data = await bffFetch.get<unknown>('/my-reservations', query);
@@ -48,4 +52,19 @@ export const createMyReservationReview = async (
   );
   if (!result) return null;
   return CreateMyReservationReviewResponseSchema.parse(result);
+};
+
+/** 예약 가능일 조회 */
+export const getAvailableSchedule = async (activityId: number, year: string, month: string) => {
+  const data = await get(
+    `/activities/${activityId}/available-schedule?year=${year}&month=${month}`
+  );
+  return z.array(ActivityScheduleSchema).parse(data);
+};
+
+export const createReservation = async (
+  activityId: number,
+  body: CreateActivityReservationRequestBody
+) => {
+  return bffFetch.post(`/activities/${activityId}/reservations`, body);
 };
